Implenent list_sources and list_movie method
